OpenVLA/requirements.txt
# OpenVLA fine-tuning dependencies
# Install via: pip install -r requirements.txt
#
# PyTorch is handled separately in setup.sh (CUDA-version-specific)

# ── Core ML ───────────────────────────────────────────────────────────────────
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.30.0
datasets>=2.18.0
bitsandbytes>=0.43.0          # quantization (optional but useful)

# ── OpenVLA model (HuggingFace hub) ──────────────────────────────────────────
huggingface_hub>=0.22.0
tokenizers>=0.19.0
safetensors>=0.4.0

# ── Data loading ──────────────────────────────────────────────────────────────
pyarrow>=14.0.0               # parquet reading
pandas>=2.0.0
decord>=0.6.0                 # fast video frame reading (mp4)
Pillow>=10.0.0
PyYAML>=6.0

# ── Experiment tracking (optional) ───────────────────────────────────────────
wandb>=0.16.0

# ── Training utilities ────────────────────────────────────────────────────────
tqdm>=4.66.0
numpy>=1.26.0
